Converted the game to webpack
